requirements.txt
flask
flask-cors
google-cloud-firestore